requirements.txt
# Requirements para o PDF Manager
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.29.7
python-dotenv==1.0.0
pydantic==2.5.0
